Fix the self-extracting WinRAR archive on Win11
When the screen's DPI is large enough the installer attempts to increase the text's size by 150%. It does this using OLECMDID_OPTICAL_ZOOM - which seems to break pre-IE9 web browser control rendering on Win11.

This commit patches the generated self-extracting archive in order to insert a <meta> tag into the html that forces the web browser control to use "modern" rendering.